# docs/09-api/openapi.yaml
openapi: 3.1.0
info:
  title: Lumine API
  description: Institutional AI-native quantitative hedge fund API
  version: 0.1.0
paths:
...
  /api/v1/portfolio/{portfolio_id}/orders:
    delete:
      tags:
      - portfolio
      summary: Cancel All Orders
      description: Cancel every open order in the portfolio (B-06 cancel-all).
      operationId: cancel_all_orders_api_v1_portfolio__portfolio_id__orders_delete
      parameters:
      - name: portfolio_id
        in: path
        required: true
        schema:
          type: string
          title: Portfolio Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelAllResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/orders:
    get:
      tags:
      - orders
      summary: List Orders
      description: List orders (DB-backed when DEMO_DATA=0).
      operationId: list_orders_api_v1_orders_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_Order_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - orders
      summary: Create Order
      description: Submit a new order (persisted when DEMO_DATA=0).
      operationId: create_order_api_v1_orders_post
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{order_id}:
    get:
      tags:
      - orders
      summary: Get Order
      description: Return a single order (DB-backed when DEMO_DATA=0).
      operationId: get_order_api_v1_orders__order_id__get
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - orders
      summary: Modify Order
      description: Modify price/volume of a pending order (ModifyOrderDialog contract).
      operationId: modify_order_api_v1_orders__order_id__patch
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyOrderRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - orders
      summary: Cancel Order
      description: Cancel an open order (persisted when DEMO_DATA=0).
      operationId: cancel_order_api_v1_orders__order_id__delete
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{order_id}/history:
    get:
      tags:
      - orders
      summary: Order History
      description: 'Append-only state transitions for one order (B-06 order history).


        Demo path yields a deterministic lifecycle (submitted → pending →

        filled); the DB path reads ``order_state_transitions`` when storage

        wiring lands.'
      operationId: order_history_api_v1_orders__order_id__history_get
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Order Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedList_OrderHistoryEntry_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/bulk-status:
    post:
      tags:
      - orders
      summary: Bulk Order Status
      description: Batch order status check (B-06).
      operationId: bulk_order_status_api_v1_orders_bulk_status_post
      parameters:
      - name: X-Lumine-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Api-Key
      - name: X-Lumine-Timestamp
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Timestamp
      - name: X-Lumine-Signature
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Lumine-Signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkStatusRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkStatusResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BulkStatusRequest:
      properties:
        order_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 100
          minItems: 1
          title: Order Ids
      type: object
      required:
      - order_ids
      title: BulkStatusRequest
      description: Batch order status check payload.
    BulkStatusResult:
      properties:
        statuses:
          additionalProperties:
            type: string
          type: object
          title: Statuses
        total:
          type: integer
          title: Total
      type: object
      required:
      - statuses
      - total
      title: BulkStatusResult
      description: Batch order status check result.
    CancelAllResult:
      properties:
        cancelled:
          type: integer
          title: Cancelled
        portfolio_id:
          type: string
          title: Portfolio Id
      type: object
      required:
      - cancelled
      - portfolio_id
      title: CancelAllResult
      description: Cancel-all outcome.
...
    CreateOrderRequest:
      properties:
        portfolio_id:
          type: string
          title: Portfolio Id
        symbol:
          type: string
          title: Symbol
          default: XAUUSD
        side:
          type: string
          enum:
          - buy
          - sell
          title: Side
        order_type:
          type: string
          enum:
          - market
          - limit
          - stop
          title: Order Type
          default: market
        volume:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Volume
        price:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Price
        stop_loss:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Stop Loss
        take_profit:
          anyOf:
          - type: number
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Take Profit
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
      type: object
      required:
      - portfolio_id
      - side
      - volume
      title: CreateOrderRequest
      description: Payload to submit a new order.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModifyOrderRequest:
      properties:
        price:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Price
        volume:
          anyOf:
          - type: number
            exclusiveMinimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Volume
      type: object
      title: ModifyOrderRequest
      description: Payload to modify a pending order (at least one field required).
    Order:
      properties:
        order_id:
          type: string
          format: uuid
          title: Order Id
        portfolio_id:
          type: string
          title: Portfolio Id
        symbol:
          type: string
          title: Symbol
        side:
          type: string
          enum:
          - buy
          - sell
          title: Side
        order_type:
          type: string
          enum:
          - market
          - limit
          - stop
          title: Order Type
        volume:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Volume
        price:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Price
        status:
          type: string
          enum:
          - pending
          - filled
          - partially_filled
          - rejected
          - cancelled
          title: Status
        filled_volume:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Filled Volume
        rejected_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Rejected Reason
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - order_id
      - portfolio_id
      - symbol
      - side
      - order_type
      - volume
      - status
      - filled_volume
      - created_at
      - updated_at
      title: Order
      description: Order lifecycle record.
    OrderHistoryEntry:
      properties:
        order_id:
          type: string
          format: uuid
          title: Order Id
        previous_state:
          type: string
          title: Previous State
        new_state:
          type: string
          title: New State
        actor_role:
          type: string
          title: Actor Role
        actor_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Actor Id
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
        decision_ts:
          type: string
          format: date-time
          title: Decision Ts
      type: object
      required:
      - order_id
      - previous_state
      - new_state
      - actor_role
      - decision_ts
      title: OrderHistoryEntry
      description: One order state transition (append-only audit).
...
    PaginatedList_OrderHistoryEntry_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/OrderHistoryEntry'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PaginatedList[OrderHistoryEntry]
    PaginatedList_Order_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/Order'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: PaginatedList[Order]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...